Refactor RAC for the inclusion of multiple modules
Before the project keeps expanding I think it is a good time to separate
into a modular architecture that allows us to expand into well defined
modules instead of having a monolithic dependency, since the RAC Host
adn the RAC Core have different requirements and run on different
environements

For this, there have been created 3 new worspaces, `rac-core`,
`rac-protocol' and 'rac-host'. The new workspaces of cargo will need
different toolchains, so this separates the Host and the Target testing
and gets everything ready to check on the CI. The testing is not
possible on RAC Core or RAC Protocol due to the `no_std` requirement so
the tests are on RAC Host for now. But this new structure will allow for
the tests to be separated for another workspace if needed.
