+++ Cargo.toml
@@ -1,13 +1,8 @@
-[package]
-name = "robot_articulation_control"
-version = "0.1.0"
-edition = "2021"
+[workspace]
+members = [
+    "rac-core",
+    "rac-protocol",
+    "rac-host"
+]
 
-[dependencies]
-bitflags = "2.6.0"
-heapless = "0.8.0"
-nalgebra = "0.33.2"
-
-[lib]
-name = "robot_articulation_control"
-path = "src/lib.rs"
+resolver = "2"
