+++ rac-core/src/encoder.rs
