+++ rac-host/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "rac-host"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+nalgebra = "0.33"
+rac-protocol = { path = "../rac-protocol" }
+rac-core = { path = "../rac-core" }
