+++ rac-protocol/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "rac-protocol"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+bitflags = "2.6.0"
+
+[features]
+default = ["no_std"]
+no_std = []
