+++ rac-protocol/src/lib.rs
@@ -1,0 +1,3 @@
+#![no_std]
+
+pub mod knode_protocol;
